Fixing fft benchmark warnings; see #114

+++ omp-tasks/fft/fft.h
@@ -2,7 +2,7 @@
 #define FFT_H
 
 /* our real numbers */
-typedef float REAL;
+typedef double REAL;
 
 /* Complex numbers and operations */
 typedef struct {
